% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraAnno.edgR.R
\name{DraAnno.edgeR}
\alias{DraAnno.edgeR}
\title{A DraAnno function for Deinococcus radiodurans annotation}
\usage{
DraAnno.edgeR(dataset, select, ..., write = T)
}
\description{
DraAnno.edgeR takes 2 argument: the name of the edgeR processed csv file(containing "logFC","logCPM","LR","PValue","FDR" variables), and in second argument, select "up" and "down" to get either up regulation or down regulation (cutoff parameters are fixed at FDR < 0.05, logFC <= -1 or >=1). Default of write is TRUE, so it automatically writes an annotion term added csv file to your current directory. You can defined the csv file name in "...".
}
\examples{
\dontrun{DraAnno.edgeR(drad_24_diff_edgeR, "up", "up24hrannotation")}




}

